src/main/java/ec/com/students/sofka/api/usecases/SaveStudentUseCase.java
package ec.com.students.sofka.api.usecases;

import ec.com.students.sofka.api.domain.collection.Student;
import ec.com.students.sofka.api.domain.dto.StudentDTO;
import ec.com.students.sofka.api.repository.IStudentRepository;
import ec.com.students.sofka.api.usecases.interfaces.SaveStudent;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.modelmapper.ModelMapper;
import org.springframework.stereotype.Service;
import org.springframework.validation.annotation.Validated;
import reactor.core.publisher.Mono;

@Service
@RequiredArgsConstructor
//@Validated
public class SaveStudentUseCase implements SaveStudent {
    private final IStudentRepository repository;

    private final ModelMapper mapper;

    @Override
    public Mono<StudentDTO> save(StudentDTO studentDTO) {
        return this.repository.save(mapper.map(studentDTO, Student.class))
                .switchIfEmpty(Mono.empty())
                .map(student -> mapper.map(student, StudentDTO.class));
    }
}
